--- sycls-core/src/main/java/com/b1/content/RoundRepository.java
@@ -1,8 +1,0 @@
-package com.b1.content;
-
-import com.b1.content.entity.Round;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoundRepository extends JpaRepository<Round, Long> {
-
-}
