+++ sycls-core/src/main/resources/application-core.yml
@@ -10,7 +10,7 @@ spring:
 
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
     properties:
       hibernate:
         show_sql: true
@@ -22,7 +22,7 @@ spring:
   sql:
     init:
       encoding: UTF-8
-      mode: embedded
+      mode: always
 
 ---
 spring:
