+++ sycls-user-api/src/main/java/com/b1/seat/SeatGradeHelper.java
@@ -1,11 +1,12 @@
 package com.b1.seat;
 
-import com.b1.content.entity.Round;
 import com.b1.exception.customexception.SeatGradeAlreadySoldOutException;
 import com.b1.exception.customexception.SeatGradeNotFoundException;
 import com.b1.exception.errorcode.SeatGradeErrorCode;
-import com.b1.seat.entity.SeatGrade;
-import com.b1.seat.entity.SeatGradeStatus;
+import com.b1.round.entity.Round;
+import com.b1.seatgrade.SeatGradeRepository;
+import com.b1.seatgrade.entity.SeatGrade;
+import com.b1.seatgrade.entity.SeatGradeStatus;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
